Rename swin model to ssi_swin and update imports
