+++ src/models/__init__.py
@@ -1,5 +1,5 @@
 """Model architectures."""
-from .swin import SSI_SwinFusionNet
+from .ssi_swin import SSI_SwinFusionNet
 from .cbam import CBAM
 
 __all__ = ["SSI_SwinFusionNet", "CBAM"]
