+++ test_model_fusion.py
@@ -1,5 +1,5 @@
 import torch
-from src.models.swin import SSI_SwinFusionNet
+from src.models.ssi_swin import SSI_SwinFusionNet
 
 model = SSI_SwinFusionNet(num_classes=4, ssi_input_dim=10, pretrained=False)
 img = torch.randn(2, 3, 224, 224)
